# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Video Analytics PPE Compliance API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CameraStatus = (typeof CameraStatus)[keyof typeof CameraStatus];

export const CameraStatus = {
  active: "active",
  inactive: "inactive",
  error: "error",
} as const;

export type CameraPpeRequirementsItem =
  (typeof CameraPpeRequirementsItem)[keyof typeof CameraPpeRequirementsItem];

export const CameraPpeRequirementsItem = {
  hard_hat: "hard_hat",
  safety_vest: "safety_vest",
  gloves: "gloves",
  safety_glasses: "safety_glasses",
  face_mask: "face_mask",
  safety_boots: "safety_boots",
} as const;

export interface Camera {
  id: number;
  name: string;
  location: string;
  status: CameraStatus;
  streamUrl?: string;
  ppeRequirements: CameraPpeRequirementsItem[];
  createdAt: string;
  updatedAt: string;
}

export type CreateCameraRequestPpeRequirementsItem =
  (typeof CreateCameraRequestPpeRequirementsItem)[keyof typeof CreateCameraRequestPpeRequirementsItem];

export const CreateCameraRequestPpeRequirementsItem = {
  hard_hat: "hard_hat",
  safety_vest: "safety_vest",
  gloves: "gloves",
  safety_glasses: "safety_glasses",
  face_mask: "face_mask",
  safety_boots: "safety_boots",
} as const;

export interface CreateCameraRequest {
  name: string;
  location: string;
  streamUrl?: string;
  ppeRequirements: CreateCameraRequestPpeRequirementsItem[];
}

export type UpdateCameraRequestStatus =
  (typeof UpdateCameraRequestStatus)[keyof typeof UpdateCameraRequestStatus];

export const UpdateCameraRequestStatus = {
  active: "active",
  inactive: "inactive",
  error: "error",
} as const;

export type UpdateCameraRequestPpeRequirementsItem =
  (typeof UpdateCameraRequestPpeRequirementsItem)[keyof typeof UpdateCameraRequestPpeRequirementsItem];

export const UpdateCameraRequestPpeRequirementsItem = {
  hard_hat: "hard_hat",
  safety_vest: "safety_vest",
  gloves: "gloves",
  safety_glasses: "safety_glasses",
  face_mask: "face_mask",
  safety_boots: "safety_boots",
} as const;

export interface UpdateCameraRequest {
  name?: string;
  location?: string;
  status?: UpdateCameraRequestStatus;
  streamUrl?: string;
  ppeRequirements?: UpdateCameraRequestPpeRequirementsItem[];
}

export type DetectedPersonPpe = {
  hard_hat: boolean;
  safety_vest: boolean;
  gloves: boolean;
  safety_glasses: boolean;
  face_mask: boolean;
  safety_boots: boolean;
};

export interface DetectedPerson {
  id: string;
  confidence: number;
  ppe: DetectedPersonPpe;
  compliant: boolean;
  missingPpe: string[];
}

export interface DetectionSnapshot {
  cameraId: number;
  timestamp: string;
  personCount: number;
  compliantCount: number;
  nonCompliantCount: number;
  complianceRate: number;
  detectedPersons: DetectedPerson[];
}

export interface LiveAnalytics {
  timestamp: string;
  totalCameras: number;
  activeCameras: number;
  totalPersonsDetected: number;
  totalCompliant: number;
  totalNonCompliant: number;
  overallComplianceRate: number;
  openAlerts: number;
  cameraSnapshots: DetectionSnapshot[];
}

export interface AnalyticsDataPoint {
  timestamp: string;
  cameraId?: number;
  personCount: number;
  compliantCount: number;
  nonCompliantCount: number;
  complianceRate: number;
}

export type ComplianceSummaryPpeMissingBreakdown = {
  hard_hat: number;
  safety_vest: number;
  gloves: number;
  safety_glasses: number;
  face_mask: number;
  safety_boots: number;
};

export type ComplianceSummaryCameraBreakdownItem = {
  cameraId: number;
  cameraName: string;
  detections: number;
  complianceRate: number;
};

export interface ComplianceSummary {
  period: string;
  totalDetections: number;
  totalCompliant: number;
  totalNonCompliant: number;
  overallComplianceRate: number;
  ppeMissingBreakdown: ComplianceSummaryPpeMissingBreakdown;
  cameraBreakdown: ComplianceSummaryCameraBreakdownItem[];
}

export type AlertType = (typeof AlertType)[keyof typeof AlertType];

export const AlertType = {
  missing_ppe: "missing_ppe",
  camera_offline: "camera_offline",
  low_compliance: "low_compliance",
} as const;

export type AlertSeverity = (typeof AlertSeverity)[keyof typeof AlertSeverity];

export const AlertSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type AlertStatus = (typeof AlertStatus)[keyof typeof AlertStatus];

export const AlertStatus = {
  open: "open",
  acknowledged: "acknowledged",
  resolved: "resolved",
} as const;

export interface Alert {
  id: number;
  cameraId: number;
  cameraName: string;
  type: AlertType;
  severity: AlertSeverity;
  message: string;
  missingPpe?: string[];
  status: AlertStatus;
  personCount: number;
  createdAt: string;
  acknowledgedAt?: string;
  resolvedAt?: string;
}

export interface AlertListResponse {
  alerts: Alert[];
  total: number;
  offset: number;
  limit: number;
}

export type DailyReportTopViolationsItem = {
  ppe: string;
  count: number;
};

export interface DailyReport {
  date: string;
  totalPersonDetections: number;
  uniqueViolations: number;
  averageComplianceRate: number;
  peakHour: string;
  peakPersonCount: number;
  alerts: number;
  hourlyData: AnalyticsDataPoint[];
  topViolations: DailyReportTopViolationsItem[];
}

export type GetAnalyticsHistoryParams = {
  cameraId?: number;
  from?: string;
  to?: string;
  interval?: GetAnalyticsHistoryInterval;
};

export type GetAnalyticsHistoryInterval =
  (typeof GetAnalyticsHistoryInterval)[keyof typeof GetAnalyticsHistoryInterval];

export const GetAnalyticsHistoryInterval = {
  minute: "minute",
  hour: "hour",
  day: "day",
} as const;

export type GetComplianceSummaryParams = {
  period?: GetComplianceSummaryPeriod;
};

export type GetComplianceSummaryPeriod =
  (typeof GetComplianceSummaryPeriod)[keyof typeof GetComplianceSummaryPeriod];

export const GetComplianceSummaryPeriod = {
  today: "today",
  week: "week",
  month: "month",
} as const;

export type ListAlertsParams = {
  cameraId?: number;
  status?: ListAlertsStatus;
  limit?: number;
  offset?: number;
};

export type ListAlertsStatus =
  (typeof ListAlertsStatus)[keyof typeof ListAlertsStatus];

export const ListAlertsStatus = {
  open: "open",
  acknowledged: "acknowledged",
  resolved: "resolved",
} as const;

export type GetDailyReportParams = {
  date?: string;
};